Update prettier to 3.x
- Updated the json-swagger plugin to support the 3.x pretter
- Switch the plugin to be JS based to avoid the need to compile via TS
- Simplified plugin to delegate most work to the base estree plugin
