+++ .prettierrc.json
@@ -8,7 +8,7 @@
       ],
       "options": {
         "plugins": [
-          "./scripts/prettier-swagger-plugin"
+          "./scripts/prettier-swagger-plugin.js"
         ],
         "parser": "json-swagger",
         "printWidth": 20,
