+++ documentation/ci-fix.md
@@ -30,18 +30,8 @@ cd specification/contosowidgetmanager
 # Install the dependencies to the local 'node_modules' folder.
 npm install
 
-# Compile TypeScript. Compilation will fail, this is expected. But it will compile 'scripts/prettier-swagger-plugin', which is what we need.
-npx tsc 
-
-# As of 5/25/2023, the prettier version should be 2.1.2
-npx prettier --version
-
 # Run 'prettier --check' to verify the problems can be reproduced locally
 npx prettier --check **/*.json
-
-# Run 'prettier --list-different' to understand which files have problems.
-# Note: there is no way to view the exact problems without actually changing the affected files. See https://github.com/prettier/prettier/issues/6069.
-npx prettier --list-different **/*.json
 
 # Run 'prettier --write' to fix the problems.
 npx prettier --write **/*.json
