+++ package.json
@@ -7,7 +7,6 @@
     "@azure-tools/typespec-azure-resource-manager": "0.33.0",
     "@azure-tools/typespec-client-generator-core": "0.33.0",
     "@azure-tools/typespec-providerhub": "0.33.0",
-    "@types/prettier": "^2.7.2",
     "@typespec/compiler": "0.47.0",
     "@typespec/http": "0.47.0",
     "@typespec/openapi": "0.47.0",
@@ -15,7 +14,7 @@
     "@typespec/versioning": "0.47.0",
     "@azure/avocado": "^0.8.4",
     "azure-rest-api-specs-eng-tools": "file:eng/tools",
-    "prettier": "^2.8.8",
+    "prettier": "~3.0.3",
     "typescript": "~5.1.3"
   },
   "private": true
